+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 2.9.8
 ----------------
 - Fix Menu default program when not allowed in Menu.php
+- Add Students/Student.php|account_created action in Student.php
 
 Changes in 2.9.7
 ----------------