+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 2.9.9
+----------------
+- Fix Do not show checkbox or radio input when no options in Referral.php & MakeReferral.php
+
 Changes in 2.9.8
 ----------------
 - Fix Menu default program when not allowed in Menu.php